fully compatabile with MSVC now. cmake transition is complete and i will now move to zed and finish the parser

+++ src/util/containers/hash_map.c
@@ -261,9 +261,9 @@ void hash_map_rehash(HashMap* hm) {
             continue;
         }
 
-        const Hash   hash        = hm->hash(ptr_offset(keys_ptr, current * hm->header->key_size));
-        const size_t fingerprint = take_fingerprint(hash);
-        size_t       probe       = hash & mask;
+        const Hash    hash        = hm->hash(ptr_offset(keys_ptr, current * hm->header->key_size));
+        const uint8_t fingerprint = take_fingerprint(hash);
+        size_t        probe       = hash & mask;
 
         // For each bucket, rehash to an index:
         // 1) before the cursor, probed into a free slot, or
